NVA Adjustments
Changed MotRifleSquad1 and MotRifleSquad23 to MotRifleSquad12 and MotRifleSquad3, to reflect the Deputy PL having one of the the three dismounted radios in the plt.

Changed the 'Gunner' role in the Squads to the 'Commander' role so people slotting know they are in charge of the vehicle and have a a level of responsibility.